requirements.txt
torch==2.4.1
torchvision==0.19.1
python-box==7.2.0
pyyaml==6.0.2
ultralytics==8.3.15
requests==2.32.3
gdown==5.2.0
mlflow==2.17.0
dagshub
ray[tune]